+++ .idea/vcs.xml
@@ -2,7 +2,5 @@
 <project version="4">
   <component name="VcsDirectoryMappings">
     <mapping directory="$PROJECT_DIR$" vcs="Git" />
-    <mapping directory="$PROJECT_DIR$/IDEA-CAMPUS-Server" vcs="Git" />
-    <mapping directory="$PROJECT_DIR$/Idea" vcs="Git" />
   </component>
 </project>
