+++ ideac-core/src/main/java/depth/main/ideac/domain/auth/application/AuthService.java
@@ -8,15 +8,19 @@ import depth.main.ideac.domain.auth.dto.req.RefreshTokenReq;
 import depth.main.ideac.domain.auth.dto.req.SignInReq;
 import depth.main.ideac.domain.auth.dto.req.SignUpReq;
 import depth.main.ideac.domain.auth.dto.res.AuthRes;
+import depth.main.ideac.domain.mail.domain.Verify;
+import depth.main.ideac.domain.mail.domain.repository.MailRepository;
 import depth.main.ideac.domain.user.domain.Role;
 import depth.main.ideac.domain.user.domain.Status;
 import depth.main.ideac.domain.user.domain.User;
 import depth.main.ideac.domain.user.domain.repository.UserRepository;
+import depth.main.ideac.domain.user.dto.PasswordReq;
 import depth.main.ideac.global.DefaultAssert;
 import depth.main.ideac.global.error.DefaultException;
 import depth.main.ideac.global.payload.ApiResponse;
 import depth.main.ideac.global.payload.ErrorCode;
 import depth.main.ideac.global.payload.Message;
+import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.http.ResponseEntity;
@@ -26,7 +30,9 @@ import org.springframework.security.core.Authentication;
 import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 
+import java.time.LocalDateTime;
 import java.util.Optional;
 
 
@@ -41,6 +47,7 @@ public class AuthService {
 
     private final UserRepository userRepository;
     private final TokenRepository tokenRepository;
+    private final MailRepository mailRepository;
 
     // 회원가입 하기
     public ResponseEntity<?> signUp(SignUpReq signUpRequest){
@@ -183,6 +190,37 @@ public class AuthService {
                 .information(userRepository.findByEmail(email).isEmpty())
                 .message("이메일 검증 완료")
                 .build();
+        return ResponseEntity.ok(apiResponse);
+    }
+
+    @Transactional
+    public ResponseEntity<?> changePassword(@Valid PasswordReq passwordReq, String code){
+
+        //검증
+        DefaultAssert.isTrue(passwordReq.getPassword().equals(passwordReq.getRePassword()), "비밀번호가 서로 다릅니다.");
+        //만료시간 검증
+        Verify verify = mailRepository.findByCode(code);
+
+        if (verify == null){
+            throw new DefaultException(ErrorCode.INVALID_CHECK, "이미변경되었습니다.");
+        }
+
+        DefaultAssert.isTrue(verify.checkExpiration(LocalDateTime.now()), "만료되었습니다.");
+
+        Optional<User> findUser = userRepository.findByEmail(verify.getEmail());
+
+        User user = findUser.get();
+        user.updatePassWord(passwordEncoder.encode(passwordReq.getPassword()));
+
+        // 인증완료 후 삭제
+        mailRepository.delete(verify);
+
+        ApiResponse apiResponse = ApiResponse.builder()
+                .check(true)
+                .information(null)
+                .message("비밀번호를 바꾸었습니다.")
+                .build();
+
         return ResponseEntity.ok(apiResponse);
     }
 }
